splitup process_validation
